+++ src/SBP_coefficients/Mattsson2014.jl
@@ -481,13 +481,19 @@ function third_derivative_coefficients(source::Mattsson2014, order::Int, T = Flo
                                                                    T(15 // 16),
                                                                    T(1 // 8),
                                                                    T(-17 // 16),
+                                                                   T(1 // 2))),
+                         # d4
+                         DerivativeCoefficientRow{T, 1, 6}(SVector(T(1 // 16),
+                                                                   T(-5 // 8),
+                                                                   T(17 // 16),
+                                                                   T(0),
+                                                                   T(-1),
                                                                    T(1 // 2))))
         right_boundary = .-left_boundary
         upper_coef = SVector(T(-1), T(1 // 2))
         central_coef = T(0)
         lower_coef = -upper_coef
-        # left_weights = SVector(T(1//2))
-        left_weights = SVector(T(1 // 2), T(1), T(1))
+        left_weights = SVector(T(1 // 2), T(1), T(1), T(1))
         right_weights = left_weights
         left_boundary_derivatives = (
                                      # first derivative